+++ apps/www/app/handler/[...stack]/page.tsx
@@ -1,5 +1,5 @@
 import { StackHandler, StackTheme } from "@stackframe/stack";
-import { stackServerApp } from "@/app/stack";
+import { stackServerApp } from "@/lib/utils/stack";
 
 export default function Handler(props: unknown) {
   return (
