+++ apps/www/app/layout.tsx
@@ -3,7 +3,7 @@ import type { ReactNode } from "react";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
 import { StackProvider } from "@stackframe/stack";
-import { stackServerApp } from "./stack";
+import { stackServerApp } from "../lib/utils/stack";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
